+++ package.json
@@ -6,11 +6,13 @@
     "@reduxjs/toolkit": "^1.8.6",
     "axios": "^0.27.2",
     "history": "^5.3.0",
+    "http-status-codes": "^2.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.4",
     "react-router-dom": "^6.4.2",
-    "react-scripts": "5.0.1"
+    "react-scripts": "5.0.1",
+    "react-toastify": "^9.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
