+++ src/components/app/app.tsx
@@ -1,5 +1,5 @@
 import {Route, BrowserRouter, Routes} from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+import { AppRoute } from '../../const';
 import { useAppSelector } from '../../hooks';
 import AddReview from '../../pages/add-review/add-review';
 import MoviePage from '../../pages/movie-page/movie-page';
@@ -14,7 +14,7 @@ import LoadingScreen from '../spinner/spinner';
 function App(): JSX.Element {
   const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
   const isFilmsDataLoading = useAppSelector((state) => state.isFilmsDataLoading);
-  if (authorizationStatus === AuthorizationStatus.Auth || isFilmsDataLoading) {
+  if (isFilmsDataLoading) {
     return (
       <LoadingScreen />
     );
