+++ src/components/tab-reviews/tab-reviews.tsx
@@ -1,4 +1,3 @@
-
 function TabReviews(): JSX.Element {
   return(
     <div className="film-card__reviews film-card__row">
