+++ src/pages/welcome-screen/welcome-screen.tsx
@@ -1,13 +1,9 @@
-/* eslint-disable indent */
 import React from 'react';
-import { Link } from 'react-router-dom';
 import FilmList from '../../components/film-list/film-list';
 import GenresList from '../../components/genres-list/genres-list';
-import { AuthorizationStatus } from '../../const';
-import { useAppSelector } from '../../hooks';
+import Header from '../../components/header/header';
 
 function WelcomeScreen(): JSX.Element {
-  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
 
   return(
     <React.Fragment>
@@ -17,31 +13,7 @@ function WelcomeScreen(): JSX.Element {
         </div>
 
         <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header film-card__head">
-          <div className="logo">
-            <a href={'/'} className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-          {authorizationStatus === AuthorizationStatus.Auth ?
-            <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <Link to={'/mylist'}><img src="img/avatar.jpg" alt="User avatar" width="63" height="63" /></Link>
-              </div>
-            </li>
-            <li className="user-block__item">
-              <a href={'/'} className="user-block__link">Sign out</a>
-            </li>
-            </ul>
-            :
-          <div className="user-block">
-          <Link to={'/login'} className="user-block__link">Sign in</Link>
-          </div>}
-        </header>
+        <Header />
 
         <div className="film-card__wrap">
           <div className="film-card__info">
