+++ src/services/api.ts
@@ -1,5 +1,15 @@
-import axios, {AxiosInstance, AxiosRequestConfig} from 'axios';
+import axios, {AxiosError, AxiosInstance, AxiosRequestConfig, AxiosResponse} from 'axios';
+import {StatusCodes} from 'http-status-codes';
+import { toast } from 'react-toastify';
 import { getToken } from './token';
+
+const StatusCodeMapping: Record<number, boolean> = {
+  [StatusCodes.BAD_REQUEST]: true,
+  [StatusCodes.UNAUTHORIZED]: true,
+  [StatusCodes.NOT_FOUND]: true
+};
+
+const shouldDisplayError = (response: AxiosResponse) => !!StatusCodeMapping[response.status];
 
 const BACKEND_URL = 'https://11.react.pages.academy/wtw';
 const REQUEST_TIMEOUT = 5000;
@@ -22,5 +32,15 @@ export const createAPI = (): AxiosInstance => {
     },
   );
 
+  api.interceptors.response.use(
+    (response) => response,
+    (error: AxiosError<{error: string}>) => {
+      if (error.response && shouldDisplayError(error.response)) {
+        toast.warn(error.response.data.error);
+      }
+
+      throw error;
+    }
+  );
   return api;
 };
